Passthrough `input` on `onUpdate` callback
